requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
attrs==21.3.0
autopep8==1.6.0
Babel==2.9.1
beautifulsoup4==4.10.0
black==21.12b0
bleach==4.1.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
cssselect==1.1.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-3-jet==1.0.8
django-allauth==0.47.0
django-summernote==0.8.20.0
django-taggit==2.0.0
django-widget-tweaks==1.4.9
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
lxml==4.7.1
markdown2==2.4.2
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.1
pluggy==1.0.0
psycopg2==2.9.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.6
pyquery==1.4.3
pytest==7.0.0rc1
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2021.3
readtime==1.1.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.0
typing_extensions==4.0.1
urllib3==1.26.7
webencodings==0.5.1
whitenoise==5.3.0
